refactor: replace constant state imports with utility functions for loop and autoplay commands
