+++ src/commands/music/play.ts
@@ -31,8 +31,8 @@ const options = {
             const { client, member, guildId } = interaction;
 
             if (!guildId) return;
-            const { searchEngine } = await client.database.getPlayer(guildId);
 
+            const { searchEngine } = await client.database.getPlayer(guildId);
             const { messages } = client.t(await client.database.getLocale(guildId)).get();
 
             if (!client.manager.useable)
@@ -47,7 +47,7 @@ const options = {
                     { name: messages.commands.play.autocomplete.noQuery, value: "https://open.spotify.com/track/4cOdK2wGLETKBW3PvgPWqT" },
                 ]);
 
-            const tracks = await client.manager.search(query, searchEngine);
+            const { tracks } = await client.manager.search(query, searchEngine);
             if (!tracks.length) return interaction.respond([{ name: messages.commands.play.autocomplete.noTracks, value: "noTracks" }]);
 
             await interaction.respond(
