+++ src/components/loop.ts
@@ -2,7 +2,7 @@ import { ActionRow, Button, ComponentCommand, type ComponentContext } from "seyf
 import { StelleOptions } from "#stelle/decorators";
 
 import { type APIButtonComponentWithCustomId, ButtonStyle, ComponentType } from "seyfert/lib/types/index.js";
-import { LOOP_STATE } from "#stelle/data/Constants.js";
+import { getLoopState } from "#stelle/utils/functions/utils.js";
 
 @StelleOptions({ inVoice: true, sameVoice: true, checkPlayer: true, checkNodes: true })
 export default class ToggleLoopComponent extends ComponentCommand {
@@ -22,7 +22,7 @@ export default class ToggleLoopComponent extends ComponentCommand {
 
         const { messages } = await ctx.getLocale();
 
-        await player.setRepeatMode(LOOP_STATE(player.repeatMode));
+        await player.setRepeatMode(getLoopState(player.repeatMode));
 
         //sussy code, but works
         const components = ctx.interaction.message.components[0].toJSON();
